+++ misc/label-placement/newaxis.R
@@ -12,6 +12,20 @@ par(mar = c(0, 0, 0, 2))
 options(.aqp.plotSPC.args = .args)
 
 plotSPC(x)
+
+plotSPC(x, depth.axis = FALSE)
+
+plotSPC(x, depth.axis = list())
+
+plotSPC(x, depth.axis = list(cex = 0.66))
+
+plotSPC(x, depth.axis = list(line = -1.5))
+
+plotSPC(x, depth.axis = list(style = 'compact'))
+plotSPC(x, depth.axis = list(style = 'traditional'))
+
+plotSPC(x, depth.axis = list(style = 'x'))
+
 plotSPC(x, max.depth = 100)
 plotSPC(x, max.depth = 54)
 plotSPC(x, max.depth = 33)